+++ jsonschema-module-swagger-2/src/main/java/com/github/victools/jsonschema/module/swagger2/Swagger2Module.java
@@ -303,7 +303,7 @@ public class Swagger2Module implements Module {
      */
     protected BigDecimal resolveMultipleOf(MemberScope<?, ?> member) {
         return this.getSchemaAnnotationValue(member, Schema::multipleOf, multipleOf -> multipleOf != 0)
-                .map(BigDecimal::new)
+                .map(BigDecimal::valueOf)
                 .orElse(null);
     }